+++ qlib/render.h
@@ -34,8 +34,8 @@ union Camera {
         Vector3 target;
         Vector3 up;
         float32 fov;
-        float32 yaw;
-        float32 pitch;
+        float64 yaw;
+        float64 pitch;
     };
     float32 E[12];
 };
